moved model and user validations to datastore layer
